requirements.txt
Django==4.0.4
django-location-field
Pillow
djangorestframework
django-filter
pygments

